+++ src/app/modules/evaluate/component/evaluate-dialog/evaluate-dialog.component.spec.ts
@@ -32,6 +32,9 @@ describe('EvaluateDialogComponent', () => {
                     useValue: {
                         language: Language.English,
                         item: {},
+                        settings: {
+                            evaluateCurrencyIds: []
+                        }
                     }
                 },
                 { provide: ElectronProvider, useClass: ElectronProviderFake }