+++ src/app/shared/module/material/material.module.ts
@@ -7,10 +7,10 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatInputModule } from '@angular/material/input';
 import { MatSelectModule } from '@angular/material/select';
 import { MatSlideToggleModule } from '@angular/material/slide-toggle';
+import { MatSliderModule } from '@angular/material/slider';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { MatTabsModule } from '@angular/material/tabs';
 import { AcceleratorComponent } from './component/accelerator/accelerator.component';
-
 
 @NgModule({
   imports: [
@@ -23,7 +23,8 @@ import { AcceleratorComponent } from './component/accelerator/accelerator.compon
     MatSelectModule,
     MatDividerModule,
     MatCardModule,
-    MatSlideToggleModule
+    MatSlideToggleModule,
+    MatSliderModule
   ],
   exports: [
     MatButtonModule,
@@ -36,6 +37,7 @@ import { AcceleratorComponent } from './component/accelerator/accelerator.compon
     MatDividerModule,
     MatCardModule,
     MatSlideToggleModule,
+    MatSliderModule,
     // custom
     AcceleratorComponent
   ],