+++ src/app/shared/module/poe/provider/index.ts
@@ -3,6 +3,4 @@ export * from './client-string.provider';
 export * from './currency/currencies.provider';
 export * from './currency/currency-chaos-equivalents.provider';
 export * from './leagues.provider';
-export * from './stats-description.provider';
 export * from './word.provider';
-