+++ src/app/shared/module/poe/service/index.ts
@@ -2,5 +2,5 @@ export * from './context.service';
 export * from './item/item-search-evaluate.service';
 export * from './item/item-search.service';
 export * from './item/parser/item-parser.service';
-export * from './stats-description/stats-description.service';
+export * from './stats/stats.service';
 export * from './word/word.service';